chore: reorganize test files into appropriate directories
