+++ test/lib/processOpenAPI/getSchemaUi.test.ts
@@ -1,8 +1,8 @@
 import { dereferenceSync } from '@trojs/openapi-dereference'
 import { describe, expect, it } from 'vitest'
-import { getSchemaExample } from '../../src/lib/examples/getSchemaExample'
-import { getSchemaUi } from '../../src/lib/processOpenAPI/getSchemaUi'
-import { specWithCircularRef } from '../testsConstants'
+import { getSchemaExample } from '../../../src/lib/examples/getSchemaExample'
+import { getSchemaUi } from '../../../src/lib/processOpenAPI/getSchemaUi'
+import { specWithCircularRef } from '../../testsConstants'
 
 interface FixtureTest {
   jsonSchema: any
