+++ test/lib/processOpenAPI/processOpenAPI.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest'
-import realSpec from '../../docs/public/openapi.json'
-import { processAsyncOpenAPI } from '../../src/lib/processOpenAPI/processAsyncOpenAPI'
-import { processOpenAPI } from '../../src/lib/processOpenAPI/processOpenAPI'
+import realSpec from '../../../docs/public/openapi.json'
+import { processAsyncOpenAPI } from '../../../src/lib/processOpenAPI/processAsyncOpenAPI'
+import { processOpenAPI } from '../../../src/lib/processOpenAPI/processOpenAPI'
 
 describe('processOpenAPI', () => {
   it('processes the OpenAPI spec', async () => {
